refactor: move action-types to internal
