+++ packages/smart-wallet/package.json
@@ -17,7 +17,6 @@
   },
   "devDependencies": {
     "@agoric/cosmic-proto": "^0.2.1",
-    "@agoric/cosmic-swingset": "^0.39.2",
     "@agoric/deploy-script-support": "^0.9.4",
     "@endo/captp": "^2.0.18",
     "@endo/init": "^0.5.52",
